Removes c folder
Since the library is migrating to C++ only (or at least mostly, i.e. I'm
not writing a C and a C++ version of any driver, so no need to separate
files by language), this commit removes the `c` folder and flattens the
structure. `avr` and `common` are now at the top level. The build system
has been updated to account for this change.