Reorganizes folder structure for hal and util (#17)
* Reorganizes AVR interfaces and drivers under a hal folder.
* Moves all "common" code to a util folder and organizes the includes
  by subfolders.
* Adds a drivers top-level folder for external peripheral drivers.
* Moves all hal code under a junk::hal namespace.

This reorginzation is to make clearer the boundaries of each portion
of the junk library. The hal section contains interfaces and hardware
abstractions for all on-chip peripherals (partly to get the library
ready for an ARM/STM32 implementation).

The "common" code was broken up more to ensure it is organized
correctly by functional groupings.

These changes are also looking ahead to the junkOS work.