Adds Red-Black Tree to containers
RbTree is a self balancing binary tree. It allows storage of up to
NumNodes items of type T. Items must be unique to ensure they are
found when requested.

Also, adds the Pair and KeyPair types which may be used as the items
in the tree.

Finally, updates some testing parameters to reduce testing time.