Adds binarySearch algorithm
Adds a binary search algorithm for pre-sorted arrays.
Unit tests include a fuzzy test that runs benchmarks on three
search algorithms: binarySearch, std::binary_search, and a basic
linearSearch. Each algorithm is run over arrays that range in
size from 8 items to 2^20 items. 100,000 searches are done and
the time it takes to search is averaged and printed to console.